Added file reading functionality
Now, the handIdentifier can get the card data from a file. The file must be formatted as a list of five cards with numerical rank and suit information as follows: two digits for the rank (e.g. 09, 14, 03, etc.), a comma, and a single digit for the suit. E.g. 09,1 is the nine of spades.
The pokerHand.txt file is a straight hand.